% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aic.R
\name{AIC_boxplots}
\alias{AIC_boxplots}
\title{Boxplots for the distribution of AIC for each method}
\usage{
AIC_boxplots(AIC_data, num_methods)
}
\arguments{
\item{AIC_data}{dataframe with the data to be plotted}

\item{num_methods}{integer representing the number of distribution methods}
}
\value{
AIC_boxplot; a boxplot for each method showing distribution of data
}
\description{
This function creates a boxplot of all the AIC values for each gene under
each tested distribution to aid in identifying outliers
}
