% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{analyze_SE}
\alias{analyze_SE}
\title{This function runs DE analysis on a count matrix in the se object}
\usage{
analyze_SE(se, method, conditions, assay_to_analyze)
}
\arguments{
\item{se}{summarizedExperiment}

\item{method}{DE analysis method}

\item{conditions}{covariate for DE analysis}

\item{assay_to_analyze}{assay for DE analysis}
}
\value{
DE analysis results
}
\description{
This function runs DE analysis on a count matrix in the se object
}
