% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{batch_design}
\alias{batch_design}
\title{This function allows you to make a batch design matrix}
\usage{
batch_design(se, batch, covariate)
}
\arguments{
\item{se}{summarized experiment object}

\item{batch}{string, batch variable}

\item{covariate}{string, biological covariate}
}
\value{
design table
}
\description{
This function allows you to make a batch design matrix
}
\examples{
library(scran)
se <- mockSCE()
batch_design_tibble <- batch_design(se, batch = "Mutation_Status",
                                                covariate = "Treatment")
batch_design_tibble

}
