% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batch_indicator}
\alias{batch_indicator}
\title{Batch and Condition indicator for signature data}
\format{
A data frame with 89 rows and 3 variables:
\describe{
\item{samples}{sample ID}
\item{batch}{batch}
\item{condition}{condition}
}
}
\usage{
data(batch_indicator)
}
\description{
This data consists of three batches and ten conditions
corresponding to case and control for the signature data
}
\keyword{datasets}
