% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorrection.R
\name{BatchCorrect}
\alias{BatchCorrect}
\title{This function allows you to Add batch corrected count matrix to the se object}
\usage{
BatchCorrect(
  se,
  Method,
  assaytouse,
  batch,
  group = NULL,
  covar,
  output_assay_name
)
}
\arguments{
\item{se}{SummarizeExperiment}

\item{Method}{Normalization Method}

\item{assaytouse}{Which assay use to do normalization}

\item{batch}{The batch}

\item{group}{The group variable}

\item{covar}{Covariate Matrix}

\item{output_assay_name}{name of results assay}
}
\value{
a summarized experiment object with normalized assay appended.
}
\description{
This function allows you to Add batch corrected count matrix to the se object
}
