% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{batchDesign}
\alias{batchDesign}
\title{This function allows you to make a batch design matrix}
\usage{
batchDesign(se, batch, covariate)
}
\arguments{
\item{se}{summarized experiment}

\item{batch}{batch variable}

\item{covariate}{biological covariate}
}
\value{
design table
}
\description{
This function allows you to make a batch design matrix
}
