% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batchQC.R
\name{batchQC}
\alias{batchQC}
\title{Checks for presence of batch effect and reports whether the batch needs to be adjusted}
\usage{
batchQC(data.matrix, nbatch = 3, ncond = 2, npercond = 10)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{nbatch}{Number of batches in the data}

\item{ncond}{Number of conditions in the data}

\item{npercond}{Number of samples per condition per batch in the data}
}
\value{
pca Principal Components Analysis object of the data
}
\description{
Checks for presence of batch effect and reports whether the batch needs to be adjusted
}
\examples{
batchQC(data.matrix)
batchQC(data.matrix, nbatch=3, ncond=2, npercond=10)
}

