% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchQC.R
\name{batchQC_analyze}
\alias{batchQC_analyze}
\title{Checks for presence of batch effect and reports whether the batch needs to be adjusted}
\usage{
batchQC_analyze(data.matrix, batch, mod = NULL)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch}
}
\value{
pca Principal Components Analysis object of the data
}
\description{
Checks for presence of batch effect and reports whether the batch needs to be adjusted
}
\examples{
nbatch <- 10
nperbatch <- 10
batch <- rep(1:nbatch, each=nperbatch)
batchQC_analyze(data.matrix, batch)
}

