% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circos.R
\name{batchqc_circosplot}
\alias{batchqc_circosplot}
\title{Produce Circos plot}
\usage{
batchqc_circosplot(dat, batch, AggMethod)
}
\arguments{
\item{dat}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{AggMethod}{Aggregation Method}
}
\description{
Produce Circos plot
}

