% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation.R
\name{batchqc_correlation}
\alias{batchqc_correlation}
\title{Produce correlation heatmap plot}
\usage{
batchqc_correlation(data.matrix, batch, mod = NULL)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch}
}
\description{
Produce correlation heatmap plot
}
\examples{
nbatch <- 10
nperbatch <- 10
batch <- rep(1:nbatch, each=nperbatch)
batchqc_correlation(data.matrix, batch)
}

