% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.R
\name{batchqc_ev_table}
\alias{batchqc_ev_table}
\title{Returns table with percent variation explained for specified number of genes}
\usage{
batchqc_ev_table(se, batch, condition, assay_name)
}
\arguments{
\item{se}{Summarized experiment object}

\item{assay_name}{Name of chosen assay}

\item{mod}{mod}

\item{batch_mod}{mod}
}
\value{
List of explained variation by batch and condition
}
\description{
Returns table with percent variation explained for specified number of genes
}
