% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variation.R
\name{batchqc_explained_variation}
\alias{batchqc_explained_variation}
\title{Returns a list of explained variation by batch and condition
combinations}
\usage{
batchqc_explained_variation(se, batch, condition = NULL, assay_name)
}
\arguments{
\item{se}{Summarized experiment object}

\item{batch}{Batch covariate}

\item{condition}{Condition covariate(s) of interest if desired, default is
NULL}

\item{assay_name}{Assay of choice}
}
\value{
List of explained variation by batch and condition
}
\description{
For “Individual Variation Variable/Batch Ratio” and “Residual Variation
Variable/Batch Ratio,” a ratio greater than 1 indicates that the batch has a
stronger effect than the variable of interest. When a batch effect is being
properly corrected, the variation explained by the batch variable should
decrease in the batch-corrected assay, indicating that the variation in the
data is now more likely to be associated with the condition of interest
rather than the batch. The ratio statistics should also be less than 1.
}
\examples{
library(scran)
se <- mockSCE()
batchqc_explained_variation <- BatchQC::batchqc_explained_variation(se,
                                        batch = "Mutation_Status",
                                        condition = "Treatment",
                                        assay_name = "counts")
batchqc_explained_variation

}
