% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchqc_fpvalue.R
\name{batchqc_f.pvalue}
\alias{batchqc_f.pvalue}
\title{Returns R2 values from F-test (full/reduced model)}
\usage{
batchqc_f.pvalue(se, mod, batch_mod, assay_name)
}
\arguments{
\item{se}{Summarized experiment object}

\item{mod}{mod}

\item{batch_mod}{mod}

\item{assay_name}{Name of chosen assay}
}
\value{
List of explained variation by batch and condition
}
\description{
Returns R2 values from F-test (full/reduced model)
}
