% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batchQC_filter_genes}
\alias{batchQC_filter_genes}
\title{Returns a datset after filtering genes of zero variance across 
batch and condition combinations}
\usage{
batchQC_filter_genes(data.matrix, batch, condition)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{condition}{Condition covariate of interest}
}
\value{
Filtered dataset after filtering genes of zero variance across 
batch and condition combinations
}
\description{
Returns a datset after filtering genes of zero variance across 
batch and condition combinations
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, ggstep=5, bbstep=15000, ccstep=10000, bvarstep=2, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
condition <- rep(rep(1:ncond, each=npercond), nbatch)
filtered.data <- batchQC_filter_genes(data.matrix, batch, condition)
}

