% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{batchqc_heatmap}
\alias{batchqc_heatmap}
\title{Produce heatmap plots for the given data}
\usage{
batchqc_heatmap(data.matrix, nbatch = 3, ncond = 2, npercond = 10)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{nbatch}{Number of batches to simulate}

\item{ncond}{Number of conditions to simulate}

\item{npercond}{Number of samples per condition per batch to simulate}
}
\description{
Produce heatmap plots for the given data
}
\examples{
batchqc_heatmap(data.matrix)
batchqc_heatmap(data.matrix, nbatch=5)
batchqc_heatmap(data.matrix, nbatch=3, ncond=2, npercond=10)
}

