% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batchqc_pc_explained_variation}
\alias{batchqc_pc_explained_variation}
\title{Returns explained variation for each principal components}
\usage{
batchqc_pc_explained_variation(pcs, vars, condition, batch)
}
\arguments{
\item{pcs}{Principal components in the given data}

\item{vars}{Variance of the Principal components in the given data}

\item{condition}{Condition covariate of interest}

\item{batch}{Batch covariate}
}
\value{
Explained variation table for each principal components
}
\description{
Returns explained variation for each principal components
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, ggstep=5, bbstep=15000, ccstep=10000, bvarstep=2, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
condition <- rep(rep(1:ncond, each=npercond), nbatch)
pdata <- data.frame(batch, condition)
modmatrix = model.matrix(~as.factor(condition), data=pdata)
pca <- batchqc_pca(data.matrix, batch, mod=modmatrix)
pcs <- t(data.frame(pca$x))
batchqc_pc_explained_variation(pcs, pca$sdev^2, condition, batch)
}

