% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{batchqc_pca}
\alias{batchqc_pca}
\title{Performs principal component analysis and
produces plot of the first two principal components}
\usage{
batchqc_pca(data.matrix, nbatch = 3, ncond = 2, npercond = 10)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{nbatch}{Number of batches to simulate}

\item{ncond}{Number of conditions to simulate}

\item{npercond}{Number of samples per condition per batch to simulate}
}
\value{
PCA object from principal component analysis
}
\description{
Performs principal component analysis and
produces plot of the first two principal components
}
\examples{
batchqc_pca(data.matrix)
batchqc_pca(data.matrix, nbatch=5)
batchqc_pca(data.matrix, nbatch=3, ncond=2, npercond=10)
}

