% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batchqc_pipeline.R
\name{batchqc_pipeline}
\alias{batchqc_pipeline}
\title{Checks for presence of batch effect and adjusts for it}
\usage{
batchqc_pipeline(data.matrix, nbatch = 3, ncond = 2, npercond = 10)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{nbatch}{Number of batches to simulate}

\item{ncond}{Number of conditions to simulate}

\item{npercond}{Number of samples per condition per batch to simulate}
}
\value{
clean.data Cleaned data
}
\description{
Checks for presence of batch effect and adjusts for it
}
\examples{
batchqc_pipeline(data.matrix)
batchqc_pipeline(data.matrix, nbatch=3, ncond=2, npercond=10)
}

