% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batchQC.R
\name{batchQC_report}
\alias{batchQC_report}
\title{Checks for presence of batch effect and creates a html report
with information including whether the batch needs to be adjusted}
\usage{
batchQC_report(data.matrix, batch, mod = NULL,
  report_file = "batchqc_report.html")
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch}
}
\value{
pca Principal Components Analysis object of the data
}
\description{
Checks for presence of batch effect and creates a html report
with information including whether the batch needs to be adjusted
}
\examples{
nbatch <- 10
nperbatch <- 10
batch <- rep(1:nbatch, each=nperbatch)
batchQC_report(data.matrix, batch)
}

