% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeAnalysis.R
\name{batchQC_shapeVariation}
\alias{batchQC_shapeVariation}
\title{Perform Mean and Variance batch variation analysis}
\usage{
batchQC_shapeVariation(data, groups, plot = FALSE, groupCol = NULL)
}
\arguments{
\item{data}{Given data}

\item{groups}{a character vector indicating sample group membership}

\item{plot}{Indicate whether to generate plot}

\item{groupCol}{group color}
}
\description{
Perform Mean and Variance batch variation analysis
}

