% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{batchQC_sva}
\alias{batchQC_sva}
\title{Estimate the surrogate variables using the 2 step approach proposed by 
Leek and Storey 2007}
\usage{
batchQC_sva(data.matrix, modmatrix)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{modmatrix}{Model matrix for outcome of interest and other covariates 
besides batch}
}
\value{
Surrogate variables analysis object
}
\description{
Estimate the surrogate variables using the 2 step approach proposed by 
Leek and Storey 2007
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
condition <- rep(rep(1:ncond, each=npercond), nbatch)
pdata <- data.frame(batch, condition)
modmatrix = model.matrix(~as.factor(condition), data=pdata)
batchQC_sva(data.matrix, modmatrix)
}

