% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\docType{class}
\name{BatchQCout-class}
\alias{BatchQCout-class}
\title{The BatchQC output class to output BatchQC results}
\description{
Contains all currently-supported BatchQC output data classes:
}
\details{
slots:
\describe{
   \item{batchqc_ev}{a single object of class list}
   \item{pca}{a single object of S3 class prcomp}
}
}
\seealso{
The constructor, \code{\link{BatchQCout}}
}

