% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batchtest.R
\name{batchtest}
\alias{batchtest}
\title{Performs test to check whether batch needs to be adjusted}
\usage{
batchtest(pca, nbatch = 3, ncond = 2, npercond = 10)
}
\arguments{
\item{pca}{PCA object from principal component analysis}

\item{nbatch}{Number of batches to simulate}

\item{ncond}{Number of conditions to simulate}

\item{npercond}{Number of samples per condition per batch to simulate}
}
\description{
Performs test to check whether batch needs to be adjusted
}
\examples{
batchtest(pca)
batchtest(pca, nbatch=5)
batchtest(pca, nbatch=3, ncond=2, npercond=10)
}

