% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bladder_data}
\alias{bladder_data}
\title{Bladder cancer data with 22,283 microarray gene expression data}
\format{
A data frame with 22,283 rows and 57 variables
}
\usage{
data(bladder_data)
}
\description{
This data consists of five batches and three conditions.
The columns are samples, and the rows represent
22,283 different microarray gene expression data.
}
\keyword{datasets}
