% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bladder_data_upload}
\alias{bladder_data_upload}
\title{Bladder data upload
This function uploads the Bladder data set from the bladderbatch package.
This dataset is from bladder cancer data with 22,283 different microarray
gene expression data. It has 57 bladder samples with 3 metadata variables
(batch, outcome and cancer). It contains 5 batches, 3 cancer types (cancer,
biopsy, control), and 5 outcomes (Biopsy, mTCC, sTCC-CIS, sTCC+CIS, and
Normal). Batch 1 contains only cancer, 2 has cancer and controls, 3 has only
controls, 4 contains only biopsy, and 5 contains cancer and biopsy}
\usage{
bladder_data_upload()
}
\value{
a SE object with counts data and metadata
}
\description{
Bladder data upload
This function uploads the Bladder data set from the bladderbatch package.
This dataset is from bladder cancer data with 22,283 different microarray
gene expression data. It has 57 bladder samples with 3 metadata variables
(batch, outcome and cancer). It contains 5 batches, 3 cancer types (cancer,
biopsy, control), and 5 outcomes (Biopsy, mTCC, sTCC-CIS, sTCC+CIS, and
Normal). Batch 1 contains only cancer, 2 has cancer and controls, 3 has only
controls, 4 contains only biopsy, and 5 contains cancer and biopsy
}
