% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bladder_meta}
\alias{bladder_meta}
\title{Real bladderbatch dataset}
\format{
A data frame with 57 rows and 3 variables:
\describe{
\item{samplesbladder}{sample ID}
\item{batch}{batch}
\item{condition}{condition}
}
}
\usage{
data(bladder_meta)
}
\description{
This dataset is from bladder cancer data. This dataset has 57 bladder samples
with 5 batches and 3 covariate levels (cancer, biopsy, control).
Batch 1 contains only cancer, 2 has cancer and controls, 3 has only controls,
4 contains only biopsy, and 5 contains cancer and biopsy
}
\keyword{datasets}
