% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BMI_data}
\alias{BMI_data}
\title{This function returns BMI data that comes form the data in "Comparing
tuberculosis gene signatures in malnourished individuals using the
TBSignatureProfiler" paper. Subject IDs were matched as shown on
"github.com/jessmcc22/BatchQCv2_Manuscript/blob/devel/R/subjectID_match.R"}
\usage{
BMI_data(meta)
}
\arguments{
\item{meta}{dataframe; metadata that needs to be matched to BMI}
}
\value{
dataframe provided as input with BMI info added
}
\description{
This function returns BMI data that comes form the data in "Comparing
tuberculosis gene signatures in malnourished individuals using the
TBSignatureProfiler" paper. Subject IDs were matched as shown on
"github.com/jessmcc22/BatchQCv2_Manuscript/blob/devel/R/subjectID_match.R"
}
