% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variation.R
\name{check_valid_input}
\alias{check_valid_input}
\title{Helper function to check for valid input}
\usage{
check_valid_input(se, batch, condition)
}
\arguments{
\item{se}{se object}

\item{batch}{batch}

\item{condition}{condition}
}
\value{
True/False boolean; True if all input is valid, False if invalid
}
\description{
Helper function to check for valid input
}
