% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{color_palette}
\alias{color_palette}
\title{Color palette}
\usage{
color_palette(n, first_hue = 25, last_hue = 360)
}
\arguments{
\item{n}{numeric object representing number of colors to be created}

\item{first_hue}{numeric object to set the first hue value}

\item{last_hue}{numeric object to set the final hue value}
}
\value{
color_list list of colors generated
}
\description{
This function creates the base color palette used in BatchQC
}
\examples{
library(scran)
n <- 100
color_list <- color_palette(n)
color_list

}
