% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{ComBat_correction}
\alias{ComBat_correction}
\title{ComBat Correction
This function applies ComBat correction to your summarized experiment object}
\usage{
ComBat_correction(se, assay_to_normalize, batch, covar, output_assay_name)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_to_normalize}{Assay that should be corrected}

\item{batch}{The variable that represents batch}

\item{covar}{list of covariates}

\item{output_assay_name}{name of results assay}
}
\value{
SE object with an added ComBat corrected array
}
\description{
ComBat Correction
This function applies ComBat correction to your summarized experiment object
}
