% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{combat_met_correction}
\alias{combat_met_correction}
\title{Combat-met Correction
This function applies combat-met correction to your summarized experiment
object}
\usage{
combat_met_correction(se, assay_to_normalize, batch, group, covar,
output_assay_name)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_to_normalize}{Assay that should be corrected}

\item{batch}{The variable that represents batch}

\item{group}{The group variable}

\item{covar}{Covariate Matrix}

\item{output_assay_name}{name of results assay}
}
\value{
SE object with an added combat-met corrected array
}
\description{
Combat-met Correction
This function applies combat-met correction to your summarized experiment
object
}
