% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{ComBat_seq_correction}
\alias{ComBat_seq_correction}
\title{ComBat-Seq Correction
This function applies ComBat-seq correction to your summarized experiment
object}
\usage{
ComBat_seq_correction(se, assay_to_normalize, batch, group, covar,
output_assay_name)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_to_normalize}{Assay that should be corrected}

\item{batch}{The variable that represents batch}

\item{group}{The group variable}

\item{covar}{list of covariates}

\item{output_assay_name}{name of results assay}
}
\value{
SE object with an added ComBat-seq corrected array
}
\description{
ComBat-Seq Correction
This function applies ComBat-seq correction to your summarized experiment
object
}
