% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batchQC.R
\name{combatPlot}
\alias{combatPlot}
\title{Adjust for batch effects using an empirical Bayes framework}
\usage{
combatPlot(dat = data.matrix, batch, mod = NULL, par.prior = TRUE,
  prior.plots = TRUE)
}
\arguments{
\item{dat}{Genomic measure matrix (dimensions probe x sample) - for example, expression matrix}

\item{batch}{{Batch covariate (only one batch allowed)}}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch}

\item{par.prior}{(Optional) TRUE indicates parametric adjustments will be used, FALSE indicates non-parametric adjustments will be used}

\item{prior.plots}{(Optional)TRUE give prior plots with black as a kernel estimate of the empirical batch effect density and red as the parametric}
}
\value{
data A probe x sample genomic measure matrix, adjusted for batch effects.
}
\description{
ComBat allows users to adjust for batch effects in datasets where the batch covariate is known, using methodology
described in Johnson et al. 2007. It uses either parametric or non-parametric empirical Bayes frameworks for adjusting data for
batch effects.  Users are returned an expression matrix that has been corrected for batch effects. The input
data are assumed to be cleaned and normalized before batch effect removal.
}

