% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{confound_metrics}
\alias{confound_metrics}
\title{Combine std. Pearson correlation coefficient and Cramer's V}
\usage{
confound_metrics(se, batch)
}
\arguments{
\item{se}{summarized experiment}

\item{batch}{batch variable}
}
\value{
metrics of confounding
}
\description{
<<<<<<< HEAD
This function combines std. pearson corr coef and cramer's V
=======
Combine std. Pearson correlation coefficient and Cramer's V
>>>>>>> 954092c10bd74f301d458186e6cd9be34fc971cb
}
