% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{confoundMetrics}
\alias{confoundMetrics}
\title{This function allows you to combine std. pearson corr coef and cramer's V}
\usage{
confoundMetrics(se, batch)
}
\arguments{
\item{se}{summarized experiment}

\item{batch}{batch variable}
}
\value{
metrics of confounding
}
\description{
This function allows you to combine std. pearson corr coef and cramer's V
}
