% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{cor_props}
\alias{cor_props}
\title{This function allows you to calculate correlation properties}
\usage{
cor_props(bd)
}
\arguments{
\item{bd}{batch design}
}
\value{
correlation properties
}
\description{
This function allows you to calculate correlation properties
}
