% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{counts2pvalue}
\alias{counts2pvalue}
\title{This function calculates p-values for each gene given counts, estimated
NB size, and estimated NB mean}
\usage{
counts2pvalue(counts, size, mu)
}
\arguments{
\item{counts}{a vector of gene expression values (in counts)}

\item{size}{an estimated size parameter of the NB distributions for the gene}

\item{mu}{a vector of estimated mu parameter of the NB distributions for
different samples of the gene}
}
\value{
a p-value based on estimated NB size and mean
}
\description{
This function calculates p-values for each gene given counts, estimated
NB size, and estimated NB mean
}
\keyword{internal}
