% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{covariate_pval_plotter}
\alias{covariate_pval_plotter}
\title{Covariate P-value Plotter
This function allows you to plot covariate pvalues of explained variation}
\usage{
covariate_pval_plotter(se, batch, condition, assay_name)
}
\arguments{
\item{se}{Summarized experiment object}

\item{batch}{Batch covariate}

\item{condition}{Condition covariate of interest}

\item{assay_name}{Assay of choice}
}
\value{
List of explained variation by batch and condition
}
\description{
Covariate P-value Plotter
This function allows you to plot covariate pvalues of explained variation
}
