% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates_not_confounded.R
\name{covariates_not_confounded}
\alias{covariates_not_confounded}
\title{Returns list of covariates not confounded by batch; helper function for
explained variation and for populating shiny app condition options}
\usage{
covariates_not_confounded(se, batch)
}
\arguments{
\item{se}{Summarized experiment object}

\item{batch}{Batch variable}
}
\value{
List of explained variation by batch and condition
}
\description{
Returns list of covariates not confounded by batch; helper function for
explained variation and for populating shiny app condition options
}
