% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{cramers_v}
\alias{cramers_v}
\title{This function allows you to calculate Cramer's V}
\usage{
cramers_v(bd)
}
\arguments{
\item{bd}{batch design}
}
\value{
Cramer's V
}
\description{
The Cramer’s V is an additional metric, similar to the Pearson Correlation,
for batch effect and will be between 0 and 1. Values closer to 0 indicate a
more balanced design and a lower likelihood of a batch effect, and values
closer to 1 are associated with an imbalanced design and a greater likelihood
of a batch effect.
}
\examples{
library(scran)
se <- mockSCE()
batch_design_tibble <- batch_design(se, batch = "Mutation_Status",
                                                covariate = "Treatment")
cramers_v_result <- BatchQC::cramers_v(batch_design_tibble)
cramers_v_result

}
