% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{DE_analyze}
\alias{DE_analyze}
\title{Differential Expression Analysis}
\usage{
DE_analyze(se, method, batch, conditions, assay_to_analyze, padj_method)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{method}{DE analysis method option
('DESeq2', 'limma', 'edgeR', 'ANOVA', or 'Kruskal-Wallis')}

\item{batch}{metadata column in the se object representing batch}

\item{conditions}{metadata columns in the se object representing additional
analysis covariates}

\item{assay_to_analyze}{Assay in the se object (either counts for DESeq2 or
normalized data for limma or edgeR) for DE analysis}

\item{padj_method}{correction method for adjusted p-value from
p.adjust.methods}
}
\value{
A named list containing the log2FoldChange, fvalue (ANOVA only),
pvalue and adjusted pvalue (padj) for each analysis returned by DESeq2,
limma, edgeR, ANOVA, or Kruskal-Wallis.
}
\description{
This function runs a user selected differential expression analysis on a
given matrix. Users shoudl ensure that the matrix meets the requirements of
the selected method. DE_Seq2 should be used for count data (non-negative,
integer values) with a negative binomial distribution, edgeR can be used for
edgeR-normalized data, ANOVA can be used for normal-distributed data,
Kruskal-Wallis test can be used as a nonparametric equivalent of one-way
ANOVA, and limma can be used for all other data (inlcuding normalized log or
log-CPM matrix. edgeR implemented in BatchQC uses glm quasi-likelihood
F-tests to test for differential expression.
}
\examples{
library(scran)
se <- mockSCE()
differential_expression <- BatchQC::DE_analyze(se = se,
                                                method = "DESeq2",
                                                batch = "Treatment",
                                                conditions = c(
                                                "Mutation_Status"),
                                                assay_to_analyze = "counts",
                                                padj_method = "BH")
pval_summary(differential_expression)
pval_plotter(differential_expression)

}
