% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram_alpha_numeric_check.R
\name{dendrogram_alpha_numeric_check}
\alias{dendrogram_alpha_numeric_check}
\title{Dendrogram alpha or numeric checker}
\usage{
dendrogram_alpha_numeric_check(dendro_category)
}
\arguments{
\item{dendro_category}{column from dendrogram object representing category}
}
\value{
geom_label label for the legend of category variable
}
\description{
This function checks if there is any numeric or strings for plotting legend
}
\examples{
library(scran)
se <- mockSCE()
dendro_alpha_numeric_check <- dendrogram_alpha_numeric_check(
                                         dendro_category = "Treatment")
dendro_alpha_numeric_check

}
