% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendrogram_plotter}
\alias{dendrogram_plotter}
\title{Dendrogram Plot}
\usage{
dendrogram_plotter(se, assay, batch_var, category_var)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay}{assay to plot}

\item{batch_var}{sample metadata column representing batch}

\item{category_var}{sample metadata column representing category of interest}
}
\value{
named list of dendrogram plots

dendrogram is a dendrogram ggplot

circular_dendrogram is a circular dendrogram ggplot
}
\description{
This function creates a dendrogram plot
}
\examples{
library(scran)

se <- mockSCE()

dendrogram_plot <- BatchQC::dendrogram_plotter(se,
                                               "counts",
                                               "Mutation_Status",
                                               "Treatment")

dendrogram_plot$dendrogram
dendrogram_plot$circular_dendrogram
}
