% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{DESeq_large_analysis}
\alias{DESeq_large_analysis}
\title{This function calculated the goodness of fit of DESeq2 for larger sample
sizes (intended for more than 150 samples).}
\usage{
DESeq_large_analysis(
    count_matrix,
    condition,
    other_variables,
    conditions_df,
    model_formula,
    num_samples,
    sampled,
    small_sample_cutoff)
}
\arguments{
\item{count_matrix}{matrix containing the data to be analyzed}

\item{condition}{a vector containing a factor of the condition of interest
(typically batch)}

\item{other_variables}{a vector of strings of other variables of interest}

\item{conditions_df}{data frame containing information for the other
variables of interest (columns in order of the other_variables vector)}

\item{model_formula}{the stat formula to be used in the DESeq analysis}

\item{num_samples}{total number of samples to analyze}

\item{sampled}{the down sampled matrix}

\item{small_sample_cutoff}{value at which non-parametric test was used
(considered "large sample size") vs parametric was used (considered
"small sample size")}
}
\value{
a list containing the string recommendation
}
\description{
This function calculated the goodness of fit of DESeq2 for larger sample
sizes (intended for more than 150 samples).
}
