% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{determine_threshold}
\alias{determine_threshold}
\title{Sets the threshold based on sample size}
\usage{
determine_threshold(sample_size)
}
\arguments{
\item{sample_size}{numeric; number of samples in the data set}
}
\value{
threshold; numeric with the threshold that should be used when
comparing significant genes
}
\description{
Sets the threshold based on sample size
}
\keyword{internal}
