% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variation.R
\name{EV_plotter}
\alias{EV_plotter}
\title{This function allows you to plot explained variation}
\usage{
EV_plotter(batchqc_ev)
}
\arguments{
\item{batchqc_ev}{table of explained variation from
batchqc_explained_variation}
}
\value{
boxplot of explained variation
}
\description{
This function allows you to plot explained variation
}
\examples{
library(scran)
se <- mockSCE()
se$Mutation_Status <- as.factor(se$Mutation_Status)
se$Treatment <- as.factor(se$Treatment)
expl_var_result <- batchqc_explained_variation(se, batch = "Mutation_Status",
                            condition = "Treatment", assay_name = "counts")
EV_boxplot <- BatchQC::EV_plotter(expl_var_result[[1]])
EV_boxplot

}
