% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\docType{data}
\name{example_batchqc_data}
\alias{example_batchqc_data}
\alias{batch_indicator}
\alias{signature_data}
\title{Batch and Condition indicator for signature data captured when 
activating different growth pathway genes in human mammary epithelial cells.}
\format{
A data frame with 89 rows and 2 variables:
\describe{
    \item{V1}{Batch Indicator}
    \item{V2}{Condition (Pathway) Indicator}
}

A data frame with 18052 rows and 89 variables:
\describe{
    \item{Columns1-89}{Control and Pathway activated samples}
    \item{rows1-18052}{Genes 1-18052}
}
}
\source{
GEO accession: GSE73628

GEO accession: GSE73628
}
\usage{
batch_indicator

signature_data
}
\value{
Batch indicator object

Signature data
}
\description{
This data consists of three batches and ten different conditions 
corresponding to control and nine different pathways

This data consists of three batches and ten different conditions 
corresponding to control and nine different pathways
}
\keyword{datasets}
