% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variation.R
\name{get.res}
\alias{get.res}
\title{Helper function to get residuals}
\usage{
get.res(y, X)
}
\arguments{
\item{y}{assay}

\item{X}{model matrix design}
}
\value{
residuals
}
\description{
Helper function to get residuals
}
