% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{getShinyInput}
\alias{getShinyInput}
\title{Getter function to get the shinyInput option}
\usage{
getShinyInput()
}
\value{
shinyInput option
}
\description{
Getter function to get the shinyInput option
}
\examples{
getShinyInput()
}

