% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{getShinyInputCombat}
\alias{getShinyInputCombat}
\title{Getter function to get the shinyInputCombat option}
\usage{
getShinyInputCombat()
}
\value{
shinyInputCombat option
}
\description{
Getter function to get the shinyInputCombat option
}
\examples{
getShinyInputCombat()
}

