% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{getShinyInputOrig}
\alias{getShinyInputOrig}
\title{Getter function to get the shinyInputOrig option}
\usage{
getShinyInputOrig()
}
\value{
shinyInputOrig option
}
\description{
Getter function to get the shinyInputOrig option
}
\examples{
getShinyInputOrig()
}

