% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{getShinyInputSVA}
\alias{getShinyInputSVA}
\title{Getter function to get the shinyInputSVA option}
\usage{
getShinyInputSVA()
}
\value{
shinyInputSVA option
}
\description{
Getter function to get the shinyInputSVA option
}
\examples{
getShinyInputSVA()
}

