% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{getShinyInputSVAf}
\alias{getShinyInputSVAf}
\title{Getter function to get the shinyInputSVAf option}
\usage{
getShinyInputSVAf()
}
\value{
shinyInputSVAf option
}
\description{
Getter function to get the shinyInputSVAf option
}
\examples{
getShinyInputSVAf()
}

