% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{getShinyInputSVAr}
\alias{getShinyInputSVAr}
\title{Getter function to get the shinyInputSVAr option}
\usage{
getShinyInputSVAr()
}
\value{
shinyInputSVAr option
}
\description{
Getter function to get the shinyInputSVAr option
}
\examples{
getShinyInputSVAr()
}

