% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeAnalysis.R
\name{gnormalize}
\alias{gnormalize}
\title{Perform Genewise Normalization of the given data matrix}
\usage{
gnormalize(dat)
}
\arguments{
\item{dat}{Given data matrix}
}
\value{
gnormdata Genewise Normalized data matrix
}
\description{
Perform Genewise Normalization of the given data matrix
}
\examples{
dat <- matrix(1:10, 2)
gnormdata <- gnormalize(dat)
}

