% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{Harman_correction}
\alias{Harman_correction}
\title{Harman Correction
This function applies Harman correction to a summarized experiment object
with and reconstructs the data back into the original feature space}
\usage{
Harman_correction(
    se,
    assay_to_normalize,
    batch,
    covar,
    output_assay_name,
    limit = 0.95,
    numrepeats = 100000L)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_to_normalize}{string; name of assay that should be corrected}

\item{batch}{factor; The variable that represents batch}

\item{covar}{string; name of covariate.}

\item{output_assay_name}{string; name of results assay}

\item{limit}{numeric; default: 0.95 Indicates the limit of confidence in
which to stop removing a batch effect. Must be between 0 and 1}

\item{numrepeats}{integer; default: 100000L the number of repeats in which
to run the simulated batch mean distribution estimator using the random
selection algorithm.}
}
\value{
SE object with an added Harman corrected reconstructed data
}
\description{
Harman Correction
This function applies Harman correction to a summarized experiment object
with and reconstructs the data back into the original feature space
}
