% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_color_palette.R
\name{heatmap_color_palette}
\alias{heatmap_color_palette}
\title{Heatmap color palette}
\usage{
heatmap_color_palette(coldata)
}
\arguments{
\item{coldata}{coldata object representing column data of heatmap}
}
\value{
col_color vector of colors corresponding to col variable
}
\description{
This function creates the color palette used in the heatmap plot
}
\examples{
library(scran)
se <- mockSCE()
heatmaps <- BatchQC::heatmap_plotter(se,
                                assay = "counts",
                                nfeature = 15,
                                annotation_column = c("Mutation_Status",
                                "Treatment"))
coldata <- heatmaps$coldata
ann_color <- BatchQC::heatmap_color_palette(coldata = coldata)
ann_color

}
