% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_num_to_char_convertor.R
\name{heatmap_num_to_char_converter}
\alias{heatmap_num_to_char_converter}
\title{Heatmap numeric to character converter}
\usage{
heatmap_num_to_char_converter(ann_col)
}
\arguments{
\item{ann_col}{column data of heatmap}
}
\value{
ann_col modified column data of heatmap
}
\description{
This function converts any found numerics to characters
}
\examples{
library(scran)
se <- mockSCE()
col_info <- colData(se)
ann_col <- heatmap_num_to_char_converter(ann_col = col_info)
ann_col

}
