% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ingest_data}
\alias{ingest_data}
\title{This function allows you to ingest data}
\usage{
ingest_data(se, group, batch)
}
\arguments{
\item{se}{summarized experiment}

\item{group}{biological covariate}

\item{batch}{batch}
}
\value{
a summarized experiment object
}
\description{
This function allows you to ingest data
}
