% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{limma_correction}
\alias{limma_correction}
\title{Limma Correction
This function applies limma batch correction to your provided assay}
\usage{
limma_correction(se, assay_to_normalize, batch, covar, output_assay_name)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_to_normalize}{Log assay that should be corrected}

\item{batch}{Factor containing batch information}

\item{covar}{list of covariates}

\item{output_assay_name}{name of results assay}
}
\value{
SE object with an added limma corrected array
}
\description{
Limma Correction
This function applies limma batch correction to your provided assay
}
