% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmfitC.R
\name{lmFitC}
\alias{lmFitC}
\title{Fit linear model for each gene given a series of arrays. 
This is the standard lmFit function from limma package with the modification
to accept an additional correlation matrix parameter option}
\usage{
lmFitC(
  object,
  design = NULL,
  ndups = 1,
  spacing = 1,
  block = NULL,
  correlation,
  cormatrix = NULL,
  weights = NULL,
  method = "ls",
  ...
)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or 
log-expression values for a series of arrays, with rows corresponding to 
genes and columns to samples. Any type of data object that can be processed 
by getEAWP is acceptable.}

\item{design}{the design matrix of the microarray experiment, with rows 
corresponding to arrays and columns to coefficients to be estimated. 
Defaults to the unit vector meaning that the arrays are treated as replicates}

\item{ndups}{positive integer giving the number of times each distinct probe 
is printed on each array.}

\item{spacing}{positive integer giving the spacing between duplicate 
occurrences of the same probe, spacing=1 for consecutive rows.}

\item{block}{vector or factor specifying a blocking variable on the arrays. 
Has length equal to the number of arrays. Must be NULL if ndups>2.}

\item{correlation}{the inter-duplicate or inter-technical replicate 
correlation}

\item{cormatrix}{the complete correlation matrix of the samples}

\item{weights}{non-negative observation weights. Can be a numeric matrix of 
individual weights, of same size as the object expression matrix, or a 
numeric vector of array weights with length equal to ncol of the expression 
matrix, or a numeric vector of gene weights with length equal to nrow of the 
expression matrix.}

\item{method}{fitting method; "ls" for least squares or "robust" for robust 
regression}

\item{...}{other optional arguments to be passed to lm.series, gls.series or 
mrlm}
}
\value{
list containing log2(quantile counts per mil reads) and library sizes
}
\description{
Fit linear model for each gene given a series of arrays. 
This is the standard lmFit function from limma package with the modification
to accept an additional correlation matrix parameter option
}
