% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{log2CPM}
\alias{log2CPM}
\title{Compute log2(counts per mil reads) and library size for each sample}
\usage{
log2CPM(qcounts, lib.size = NULL)
}
\arguments{
\item{qcounts}{quantile normalized counts}

\item{lib.size}{default is colsums(qcounts)}
}
\value{
list containing log2(quantile counts per mil reads) and library sizes
}
\description{
Compute log2(counts per mil reads) and library size for each sample
}

