% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{makeSVD}
\alias{makeSVD}
\title{Compute singular value decomposition}
\usage{
makeSVD(x)
}
\arguments{
\item{x}{matrix of genes by sample (ie. the usual data matrix)}
}
\value{
returns a list of svd components v and d
}
\description{
Compute singular value decomposition
}

