% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_IDs}
\alias{merged_IDs}
\title{BMI and matched sample names for TB data}
\format{
A data frame with 91 rows and 3 columns
\describe{
\item{subjectID_curatedTBData}{Subject ID found in curatedTBData}
\item{subjectID_TB_Paper}{Subject ID in the original paper}
\item{BMI}{subject's BMI from the original study}
}
}
\usage{
data(merged_IDs)
}
\description{
This is support data for the TB data set that contains the BMI data and ID
numbers from both the curatedTBData database and the original study the data
was used in
}
\keyword{datasets}
