% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methylation_data}
\alias{methylation_data}
\title{DNA methylation data with 100 methylation levels}
\format{
A data frame with 100 rows and 20 variables
}
\usage{
data(methylation_data)
}
\description{
This data consists of two batches and two conditions
corresponding to case and control. The columns are case/control
samples, and the rows represent 100 different sites.
}
\keyword{datasets}
