% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methylation_sample_info}
\alias{methylation_sample_info}
\title{Batch and Condition indicator for DNA methylation data}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{batch}{Batch Indicator}
\item{category}{Condition (Case vs Control) Indicator}
}
}
\usage{
data(methylation_sample_info)
}
\description{
This data consists of two batches and two conditions
corresponding to case and control for the DNA methylation data
}
\keyword{datasets}
