% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{mvalue_combat_correction}
\alias{mvalue_combat_correction}
\title{Mvalue-ComBat Correction
This function applies Mvalue-ComBat correction to your summarized experiment
object}
\usage{
mvalue_combat_correction(se, assay_to_normalize, batch, group, covar,
output_assay_name)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_to_normalize}{Assay that should be corrected}

\item{batch}{The variable that represents batch}

\item{group}{The group variable}

\item{covar}{Covariate Matrix}

\item{output_assay_name}{name of results assay}
}
\value{
SE object with an added mvalue-combat corrected array
}
\description{
Mvalue-ComBat Correction
This function applies Mvalue-ComBat correction to your summarized experiment
object
}
