% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{nb_down_sample}
\alias{nb_down_sample}
\title{Performs down sampling for negative binomail model fit check.}
\usage{
nb_down_sample(count_matrix, num_genes)
}
\arguments{
\item{count_matrix}{matrix; contains the feature data (should be counts)}

\item{num_genes}{integer; number of genes to use in down sampling}
}
\value{
list containing "sampled" a down sampled assay and "count_matrix"
with the down sampled count_marix
}
\description{
Performs down sampling for negative binomail model fit check.
}
\keyword{internal}
