% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{NormalizateSE}
\alias{NormalizateSE}
\title{This function allows you to Add normalized count matrix to the se object}
\usage{
NormalizateSE(se, Method, assaytouse, output_assay_name)
}
\arguments{
\item{se}{SummarizeExperiment}

\item{Method}{Normalization Method}

\item{assaytouse}{Which assay use to do normalization}

\item{output_assay_name}{name of results assay}
}
\value{
a summarized experiment object with normalized assay appended.
}
\description{
This function allows you to Add normalized count matrix to the se object
}
