% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{PCA_preprocess}
\alias{PCA_preprocess}
\title{Preprocess normalized count data for PCA}
\usage{
PCA_preprocess(se, assay, nfeature)
}
\arguments{
\item{se}{Summarized Experiment object}

\item{assay}{Assay from SummarizedExperiment object}

\item{nfeature}{Number of variable features to use}
}
\value{
Returns a list with class "prcomp" (see ?stats::prcomp)
}
\description{
Preprocess normalized count data for PCA
}
