% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_data}
\alias{plot_data}
\title{This function formats the PCA plot using ggplot}
\usage{
plot_data(pca_plot_data, color, shape, batch, xaxisPC, yaxisPC)
}
\arguments{
\item{pca_plot_data}{Data for all assays to plot}

\item{color}{variable that will be plotted as color}

\item{shape}{variable that will be plotted as shape}

\item{batch}{variable representing batch for the ellipses}

\item{xaxisPC}{the PC to plot as the x axis}

\item{yaxisPC}{the PC to plot as the y axis}
}
\value{
PCA plot
}
\description{
This function formats the PCA plot using ggplot
}
