% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_samplewise_moments}
\alias{plot_samplewise_moments}
\title{Visualize sample-wise moments}
\usage{
plot_samplewise_moments(data, batch, robust)
}
\arguments{
\item{data}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{robust}{Boolean indicator of using robust (TRUE) or non-robust test (FALSE) in visualization}
}
\value{
Sample-wise moments
}
\description{
Visualize sample-wise moments
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
condition <- rep(rep(1:ncond, each=npercond), nbatch)
data_adjusted <- batchQC_condition_adjusted(data.matrix, batch, condition)
sample_moments <- plot_samplewise_moments(data_adjusted, batch, robust=FALSE)
}
