% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotPC}
\alias{plotPC}
\title{Plot first 2 principal components}
\usage{
plotPC(v, d, ...)
}
\arguments{
\item{v}{from makeSVD}

\item{d}{from makeSVD}

\item{...}{pass options to internal plot fct.}
}
\value{
a plot
}
\description{
Plot first 2 principal components
}

