% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{possible_distances}
\alias{possible_distances}
\title{Create potential min_distance values for exploratory analysis based on
the value of spread}
\usage{
possible_distances(spread)
}
\arguments{
\item{spread}{numeric; the value of spread used in the exploratory analysis}
}
\value{
vector of min_distance values to use in exploratory analysis
}
\description{
Create potential min_distance values for exploratory analysis based on
the value of spread
}
