% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{possible_k_neighbors}
\alias{possible_k_neighbors}
\title{Create a vector of possible nearest neighbor values from 5, 15, 25, 50,
and 100}
\usage{
possible_k_neighbors(data_size)
}
\arguments{
\item{data_size}{size of the data set used to create umaps}
}
\value{
k nearest neighbor list
}
\description{
Create a vector of possible nearest neighbor values from 5, 15, 25, 50,
and 100
}
