% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess assay data}
\usage{
preprocess(se, assay, nfeature, log_option)
}
\arguments{
\item{se}{Summarized Experiment object}

\item{assay}{Assay from SummarizedExperiment object}

\item{nfeature}{Number of variable features to use}

\item{log_option}{"True" if data should be logged, "False" otherwise}
}
\value{
Returns processed data
}
\description{
Preprocess assay data
}
