% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{process_dendrogram}
\alias{process_dendrogram}
\title{Process Dendrogram}
\usage{
process_dendrogram(se, assay, batch_var)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay}{assay to plot}

\item{batch_var}{sample metadata column}
}
\value{
named list of dendrogram data

dendrogram_segments is data representing segments of the dendrogram

dendrogram_ends is data representing ends of the dendrogram
}
\description{
This function processes count data for dendrogram plotting
}
