% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_data}
\alias{protein_data}
\title{Protein data with 39 protein expression levels}
\format{
A data frame with 39 rows and 24 variables
}
\usage{
protein_data
}
\description{
This data consists of two batches and two conditions
corresponding to case and control. The columns are case/control
samples, and the rows represent 39 different proteins.
}
\examples{
data(protein_data)
}
\keyword{datasets}
