% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aic.R
\name{run_AIC_models}
\alias{run_AIC_models}
\title{Helper function that contains the code to run the lognormal, voom, and
negative binomial AIC models for \code{compute_aic}}
\usage{
run_AIC_models(dat, design, nb_test, maxit)
}
\arguments{
\item{dat}{dataframe of the data to analyze}

\item{design}{stats design model to be used in the analyses}

\item{nb_test}{boolean; should negative binomial run (must be discrete data)}

\item{maxit}{integer; the max number of IWLS iterations}
}
\value{
data frame; containing the AIC results of each method
}
\description{
Helper function that contains the code to run the lognormal, voom, and
negative binomial AIC models for \code{compute_aic}
}
