% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{setShinyInput}
\alias{setShinyInput}
\title{Setter function to set the shinyInput option}
\usage{
setShinyInput(x)
}
\arguments{
\item{x}{shinyInput option}
}
\value{
shinyInput option
}
\description{
Setter function to set the shinyInput option
}
\examples{
setShinyInput(NULL)
}

