% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{setShinyInputCombat}
\alias{setShinyInputCombat}
\title{Setter function to set the shinyInputCombat option}
\usage{
setShinyInputCombat(x)
}
\arguments{
\item{x}{shinyInputCombat option}
}
\value{
shinyInputCombat option
}
\description{
Setter function to set the shinyInputCombat option
}
\examples{
setShinyInputCombat(NULL)
}

