% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{setShinyInputOrig}
\alias{setShinyInputOrig}
\title{Setter function to set the shinyInputOrig option}
\usage{
setShinyInputOrig(x)
}
\arguments{
\item{x}{shinyInputOrig option}
}
\value{
shinyInputOrig option
}
\description{
Setter function to set the shinyInputOrig option
}
\examples{
setShinyInputOrig(NULL)
}

