% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{setShinyInputSVA}
\alias{setShinyInputSVA}
\title{Setter function to set the shinyInputSVA option}
\usage{
setShinyInputSVA(x)
}
\arguments{
\item{x}{shinyInputSVA option}
}
\value{
shinyInputSVA option
}
\description{
Setter function to set the shinyInputSVA option
}
\examples{
setShinyInputSVA(NULL)
}

