% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{setShinyInputSVAf}
\alias{setShinyInputSVAf}
\title{Setter function to set the shinyInputSVAf option}
\usage{
setShinyInputSVAf(x)
}
\arguments{
\item{x}{shinyInputSVAf option}
}
\value{
shinyInputSVAf option
}
\description{
Setter function to set the shinyInputSVAf option
}
\examples{
setShinyInputSVAf(NULL)
}

