% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{setShinyInputSVAr}
\alias{setShinyInputSVAr}
\title{Setter function to set the shinyInputSVAr option}
\usage{
setShinyInputSVAr(x)
}
\arguments{
\item{x}{shinyInputSVAr option}
}
\value{
shinyInputSVAr option
}
\description{
Setter function to set the shinyInputSVAr option
}
\examples{
setShinyInputSVAr(NULL)
}

