% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature_data}
\alias{signature_data}
\title{Signature data with 1600 gene expression levels}
\format{
A data frame with 1600 rows and 89 variables
}
\usage{
data(signature_data)
}
\description{
This data consists of three batches and ten conditions.
The columns are samples, and the rows represent
1600 different genes.
}
\keyword{datasets}
