% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{std_pearson_corr_coef}
\alias{std_pearson_corr_coef}
\title{Calculate a standardized Pearson correlation coefficient}
\usage{
std_pearson_corr_coef(bd)
}
\arguments{
\item{bd}{batch design}
}
\value{
standardized Pearson correlation coefficient
}
\description{
Pearson correlation coefficient indicates the strength of the linear
relationship between your batch and condition variables. Note that this is
the correlation between two categorical variables, so rather than measuring
the correlation between the two variables (as with two continuous variables),
it directly measures the dependence/independence of the batch and condition
in the experimental design—i.e., the potential for confounding, but not
confounding itself. Values for this metric can range from -1 to 1, with a
value close to 0 indicating batch/condition independence or a “balanced
design”, which should be associated with a lower risk of observing a batch
effect. A value closer to -1 or 1 indicates greater imbalance in the design
and a higher likelihood of a batch effect.
}
\examples{
library(scran)
se <- mockSCE()
batch_design_tibble <- batch_design(se, batch = "Mutation_Status",
                                                covariate = "Treatment")
pearson_cor_result <- BatchQC::std_pearson_corr_coef(batch_design_tibble)
pearson_cor_result

}
