% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{summarize_experiment}
\alias{summarize_experiment}
\title{This function allows you to create aa summarized experiment object}
\usage{
summarize_experiment(Counts_path, metadata_path)
}
\arguments{
\item{metadata_path}{path to metadata file}

\item{counts_path}{path to counts file}
}
\value{
a summarized experiment object
}
\description{
This function allows you to create aa summarized experiment object
}
