% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variation.R
\name{summary_stats_EV_table}
\alias{summary_stats_EV_table}
\title{Summary Stats EV Table
Returns table with min, 1st quartile, mean, 2nd quartile, and max for each
variable in the explained variation boxplot}
\usage{
summary_stats_EV_table(batchqc_ev)
}
\arguments{
\item{batchqc_ev}{explained variation results from}
}
\value{
summary_stats_table dataframe containing the min, 1st quartile, mean,
2nd quartile, and max for each variable in the explained variation boxplot
}
\description{
Summary Stats EV Table
Returns table with min, 1st quartile, mean, 2nd quartile, and max for each
variable in the explained variation boxplot
}
\examples{
library(scran)
se <- mockSCE()
se$Mutation_Status <- as.factor(se$Mutation_Status)
se$Treatment <- as.factor(se$Treatment)
exp_var_result <- BatchQC::batchqc_explained_variation(se,
                                    batch = "Mutation_Status",
                                    condition = "Treatment",
                                    assay_name = "counts")
summary_stats_table <- BatchQC::summary_stats_EV_table(exp_var_result[[1]])

summary_stats_table

}
