% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tb_data_upload}
\alias{tb_data_upload}
\title{TB data upload
This function uploads the TB data set from the curatedTBData package.}
\usage{
tb_data_upload()
}
\value{
a SE object with raw counts data and metadata
}
\description{
TB data upload
This function uploads the TB data set from the curatedTBData package.
}
\examples{
library(curatedTBData)
se_object <- tb_data_upload()

}
