% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{transform_SE}
\alias{transform_SE}
\title{This function allows you to add transformed count matrix to the SE object}
\usage{
transform_SE(se, method, assay_to_transform, output_assay_name)
}
\arguments{
\item{se}{SummarizedExperiment Object}

\item{method}{Transformation Method, currently logit only}

\item{assay_to_transform}{Which SE assay to do transformation on}

\item{output_assay_name}{name for the resulting transformed assay}
}
\value{
the original SE object with transformed assay appended
}
\description{
This function allows you to add transformed count matrix to the SE object
}
