% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{analyze_SE}
\alias{analyze_SE}
\title{This function allows you to run differential expression analysis on a count matrix in the se object}
\usage{
analyze_SE(se, method, conditons, assay_to_analyze)
}
\arguments{
\item{se}{SummarizeExperiment}

\item{method}{Differential Expression Analysis Method}

\item{assay_to_analyze}{Which assay to do DE analysis with}

\item{conditions}{DE analysis experimental design}
}
\value{
a data table of DE analysis results
}
\description{
This function allows you to run differential expression analysis on a count matrix in the se object
}
