% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{analyze_SE}
\alias{analyze_SE}
\title{This function allows you to Add normalized count matrix to the se object}
\usage{
analyze_SE(se, method, conditions, assay_to_analyze, output_assay_name)
}
\arguments{
\item{se}{SummarizeExperiment}

\item{method}{Differential Expression Analysis Method}

\item{conditions}{Condition covariate of interest}

\item{assay_to_analyze}{Which assay use to do DE analysis}

\item{output_assay_name}{name of results assay}
}
\value{
a summarized experiment object with normalized assay appended.
}
\description{
This function allows you to Add normalized count matrix to the se object
}
