% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{batch_correct}
\alias{batch_correct}
\title{Batch Correct
This function allows you to Add batch corrected count matrix to the SE object}
\usage{
batch_correct(
    se,
    method,
    assay_to_normalize,
    batch,
    group = NULL,
    covar,
    output_assay_name
)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{method}{Normalization Method}

\item{assay_to_normalize}{Which assay use to do normalization}

\item{batch}{The batch}

\item{group}{The group variable}

\item{covar}{Covariate Matrix}

\item{output_assay_name}{name of results assay}
}
\value{
a summarized experiment object with normalized assay appended.
}
\description{
Batch Correct
This function allows you to Add batch corrected count matrix to the SE object
}
