% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{batch_design}
\alias{batch_design}
\title{This function allows you to make a batch design matrix}
\usage{
batch_design(se, batch, covariate)
}
\arguments{
\item{se}{summarized experiment}

\item{batch}{batch variable}

\item{covariate}{biological covariate}
}
\value{
design table
}
\description{
This function allows you to make a batch design matrix
}
