% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batch_indicator}
\alias{batch_indicator}
\title{Batch and Condition indicator for signature data}
\format{
A data frame with 89 rows and 2 variables:
\describe{
\item{batch}{batch}
\item{condition}{condition}
}
}
\usage{
data(batch_indicator)
}
\description{
This dataset is from signature data captured when activating different growth
pathway genes in human mammary epithelial cells (GEO accession: GSE73628).
This data consists of three batches and ten different conditions
corresponding to control and nine different pathways.
}
\keyword{datasets}
