% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{batch_pval_plotter}
\alias{batch_pval_plotter}
\title{This function allows you to plot batch p-values of explained variation}
\usage{
batch_pval_plotter(DE_res)
}
\arguments{
\item{DE_res}{DE analysis results output from DE_analyze()}
}
\value{
List of explained variation by batch and condition
}
\description{
This function allows you to plot batch p-values of explained variation
}
