% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{batch_pval_plotter}
\alias{batch_pval_plotter}
\title{This function allows you to plot batch p-values of explained variation}
\usage{
batch_pval_plotter(dds)
}
\arguments{
\item{dds}{DESeq2 analysis results output}
}
\value{
List of explained variation by batch and condition
}
\description{
This function allows you to plot batch p-values of explained variation
}
