% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchQC.R
\name{batchQC}
\alias{batchQC}
\title{Checks for presence of batch effect and creates a html report
with information including whether the batch needs to be adjusted}
\usage{
batchQC(dat, batch, mod = NULL, report_file = "batchqc_report.html",
  report_dir = ".", report_option_binary = "111111111",
  view_report = TRUE, interactive = TRUE)
}
\arguments{
\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch}

\item{report_file}{Output report file name}

\item{report_dir}{Output report directory path}

\item{report_option_binary}{9 bits Binary String representing the plots to display and hide in the report}

\item{view_report}{when TRUE, opens the report in a browser}

\item{data.matrix}{Given data or simulated data from rnaseq_sim()}
}
\value{
pca Principal Components Analysis object of the data
}
\description{
Checks for presence of batch effect and creates a html report
with information including whether the batch needs to be adjusted
}
\examples{
nbatch <- 10
nperbatch <- 10
batch <- rep(1:nbatch, each=nperbatch)
batchQC(data.matrix, batch)
}

