% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{batchQC}
\alias{batchQC}
\title{Checks for presence of batch effect and creates a html report 
with information including whether the batch needs to be adjusted}
\usage{
batchQC(dat, batch, condition = NULL, report_file = "batchqc_report.html",
    report_dir = ".", report_option_binary = "111111111",
    view_report = TRUE, interactive = TRUE, batchqc_output = FALSE,
    log2cpm_transform = FALSE)
}
\arguments{
\item{dat}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{condition}{Covariates or conditions of interest besides batch}

\item{report_file}{Output report file name}

\item{report_dir}{Output report directory path}

\item{report_option_binary}{9 bits Binary String representing 
the plots to display and hide in the report}

\item{view_report}{when TRUE, opens the report in a browser}

\item{interactive}{when TRUE, opens the interactive shinyApp}

\item{batchqc_output}{when TRUE, creates BatchQCout object in 
batchqc_output.rda R object file}

\item{log2cpm_transform}{when TRUE, transforms the data using log2CPM -
log2 Counts Per Million transformation function}
}
\value{
outputfile Report file generated by batchQC
}
\description{
Checks for presence of batch effect and creates a html report 
with information including whether the batch needs to be adjusted
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
condition <- rep(rep(1:ncond, each=npercond), nbatch)
batchQC(data.matrix, batch=batch, condition=condition, view_report=FALSE, 
    interactive=FALSE)
}

