% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{BatchQC}
\alias{BatchQC}
\title{Run BatchQC shiny app}
\usage{
BatchQC(dev = FALSE)
}
\arguments{
\item{dev}{Run the applicaiton in developer mode}
}
\value{
The shiny app will open
}
\description{
Run BatchQC shiny app
}
\examples{
\dontrun{
BatchQC()
}
}
