% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\name{batchQC_analyze}
\alias{batchQC_analyze}
\title{Checks for presence of batch effect and reports whether the batch 
needs to be adjusted}
\usage{
batchQC_analyze(data.matrix, batch, mod = NULL)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates 
besides batch}
}
\value{
pca Principal Components Analysis object of the data
}
\description{
Checks for presence of batch effect and reports whether the batch 
needs to be adjusted
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
condition <- rep(rep(1:ncond, each=npercond), nbatch)
pdata <- data.frame(batch, condition)
modmatrix = model.matrix(~as.factor(condition), data=pdata)
batchQC_analyze(data.matrix, batch, mod=modmatrix)
}

