% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circos.R
\name{batchqc_circosplot}
\alias{batchqc_circosplot}
\title{Produce Circos plot}
\usage{
batchqc_circosplot(dat, batch, AggMethod)
}
\arguments{
\item{dat}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{AggMethod}{Aggregation Method}
}
\value{
Generates Circular Dendrogram plot for the given data
}
\description{
Produce Circos plot
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
batchqc_circosplot(data.matrix, batch=batch, AggMethod='complete')
}

