% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batchqc_explained_variation}
\alias{batchqc_explained_variation}
\title{Returns a list of explained variation by batch and condition 
combinations}
\usage{
batchqc_explained_variation(data.matrix, condition, batch)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{condition}{Condition covariate of interest}

\item{batch}{Batch covariate}
}
\value{
List of explained variation by batch and condition
}
\description{
Returns a list of explained variation by batch and condition 
combinations
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
condition <- rep(rep(1:ncond, each=npercond), nbatch)
batchqc_explained_variation(data.matrix, condition, batch)
}

