% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.R
\name{batchqc_explained_variation}
\alias{batchqc_explained_variation}
\title{Returns a list of explained variation by batch and condition
combinations}
\usage{
batchqc_explained_variation(se, condition, batch)
}
\arguments{
\item{se}{Summarized experiment object}

\item{condition}{Condition covariate of interest}

\item{batch}{Batch covariate}
}
\value{
List of explained variation by batch and condition
}
\description{
Returns a list of explained variation by batch and condition
combinations
}
