% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.R
\name{batchqc_explained_variation}
\alias{batchqc_explained_variation}
\title{Returns a list of explained variation by batch and condition
combinations}
\usage{
batchqc_explained_variation(se, batch, condition, assay_name)
}
\arguments{
\item{se}{Summarized experiment object}

\item{batch}{Batch covariate}

\item{condition}{Condition covariate of interest}

\item{assay_name}{Assay of choice}
}
\value{
List of explained variation by batch and condition
}
\description{
Returns a list of explained variation by batch and condition
combinations
}
