% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{batchqc_heatmap}
\alias{batchqc_heatmap}
\title{Produce heatmap plots for the given data}
\usage{
batchqc_heatmap(data.matrix, batch, mod = NULL)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch}
}
\description{
Produce heatmap plots for the given data
}
\examples{
nbatch <- 10
nperbatch <- 10
batch <- rep(1:nbatch, each=nperbatch)
batchqc_heatmap(data.matrix, batch)
}

