% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{batchqc_heatmap}
\alias{batchqc_heatmap}
\title{Produce heatmap plots for the given data}
\usage{
batchqc_heatmap(data.matrix, batch, mod = NULL, max_display = 50)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates 
besides batch}

\item{max_display}{Maximum number of rows to display in heat map}
}
\value{
Heatmap plots for the given data
}
\description{
Produce heatmap plots for the given data
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
condition <- rep(rep(1:ncond, each=npercond), nbatch)
pdata <- data.frame(batch, condition)
modmatrix = model.matrix(~as.factor(condition), data=pdata)
batchqc_heatmap(data.matrix, batch, mod=modmatrix)
}

