% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.R
\name{batchqc_pca}
\alias{batchqc_pca}
\title{Performs principal component analysis and
produces plot of the first two principal components}
\usage{
batchqc_pca(data.matrix, batch, mod = NULL)
}
\arguments{
\item{data.matrix}{Given data or simulated data from rnaseq_sim()}

\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch}
}
\value{
PCA object from principal component analysis
}
\description{
Performs principal component analysis and
produces plot of the first two principal components
}
\examples{
nbatch <- 10
nperbatch <- 10
batch <- rep(1:nbatch, each=nperbatch)
batchqc_pca(data.matrix, batch)
}

