% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeAnalysis.R
\name{batchQC_shapeVariation}
\alias{batchQC_shapeVariation}
\title{Perform Mean and Variance batch variation analysis}
\usage{
batchQC_shapeVariation(data, groups, plot = FALSE, groupCol = NULL)
}
\arguments{
\item{data}{Given data}

\item{groups}{a character vector indicating sample group membership}

\item{plot}{Indicate whether to generate plot}

\item{groupCol}{group color}
}
\value{
Mean and Variance batch variation Overall and Pairwise p-values
}
\description{
Perform Mean and Variance batch variation analysis
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
batchQC_shapeVariation(data.matrix, groups=batch)
}

