% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batchtest.R
\name{batchtest}
\alias{batchtest}
\title{Performs test to check whether batch needs to be adjusted}
\usage{
batchtest(pca, batch, mod = NULL)
}
\arguments{
\item{pca}{PCA object from principal component analysis}

\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch}
}
\description{
Performs test to check whether batch needs to be adjusted
}
\examples{
nbatch <- 10
nperbatch <- 10
batch <- rep(1:nbatch, each=nperbatch)
batchtest(pca, batch)
}

