% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtest.R
\name{batchtest}
\alias{batchtest}
\title{Performs test to check whether batch needs to be adjusted}
\usage{
batchtest(pca, batch, mod = NULL)
}
\arguments{
\item{pca}{PCA object from principal component analysis}

\item{batch}{Batch covariate}

\item{mod}{Model matrix for outcome of interest and other covariates 
besides batch}
}
\value{
Summary of linear regression of first five principal components
}
\description{
Performs test to check whether batch needs to be adjusted
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
batch <- rep(1:nbatch, each=ncond*npercond)
condition <- rep(rep(1:ncond, each=npercond), nbatch)
pdata <- data.frame(batch, condition)
modmatrix = model.matrix(~as.factor(condition), data=pdata)
pca <- batchqc_pca(data.matrix, batch, mod=modmatrix)
batchtest(pca, batch, mod=modmatrix)
}

