% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BMI_data}
\alias{BMI_data}
\title{This function returns BMI data that comes form the indata object provided to
Jessica from Evan}
\usage{
BMI_data()
}
\value{
a SE object with raw counts data and metadata
}
\description{
This function returns BMI data that comes form the indata object provided to
Jessica from Evan
}
