% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variation.R
\name{check_valid_input}
\alias{check_valid_input}
\title{Helper function to save variables as factors if not already factors}
\usage{
check_valid_input(se, batch, condition)
}
\arguments{
\item{se}{se object}

\item{batch}{batch}

\item{condition}{condition}
}
\value{
se se object
}
\description{
Helper function to save variables as factors if not already factors
}
\keyword{Internal}
