% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{combat_correction}
\alias{combat_correction}
\title{Combat Correction
This function applies combat correction to your summarized experiment object}
\usage{
combat_correction(se, assay_to_normalize, batch, covar, output_assay_name)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_to_normalize}{Assay that shoudl be corrected}

\item{batch}{The variable that represents batch}

\item{covar}{Covariate Matrix}

\item{output_assay_name}{name of results assay}
}
\value{
SE object with an added combat corrected array
}
\description{
Combat Correction
This function applies combat correction to your summarized experiment object
}
