% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{commentary}
\alias{commentary}
\title{This function creates the commentary recommendation when there are more than
20 samples.}
\usage{
commentary(
  nb_fit,
  nb_fit_pval,
  count_below_value,
  count_below_value_pval,
  low_pval,
  method
)
}
\arguments{
\item{nb_fit}{Boolean representing if the count is below the threshold}

\item{nb_fit_pval}{Boolean representing if the p-val count is below threshold}

\item{count_below_value}{number of features below threshold}

\item{count_below_value_pval}{number of features below p-val threshold}

\item{low_pval}{pval threshold}

\item{method}{string; method utilized for the parametric or non-parametric
test; either "DESeq2" or "edgeR"}
}
\value{
a commentary string statement
}
\description{
This function creates the commentary recommendation when there are more than
20 samples.
}
