% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aic.R
\name{compute_aic}
\alias{compute_aic}
\title{Compute the AIC for lognormal (ComBat) model, negative binomial (ComBat-seq)
model and the Voom model}
\usage{
compute_aic(se, assay_of_interest, batchind, groupind)
}
\arguments{
\item{batchind}{Factor or numeric vector of length = ncol(dat);
batch indicator for each sample.}

\item{groupind}{Factor or numeric vector of length = ncol(dat);
biological group label/indicator for each sample.}

\item{dat}{Numeric matrix of dimension (genes by samples).
Each row represents one gene's expression across samples.}
}
\value{
A list with the following two elements:
\describe{
\item{total_AIC}{The sum of AICs across all genes for the three models
in comparison.}
\item{min_AIC}{The number of minimum AIC across the three models in
comparison for individual genes.}
}
}
\description{
\describe{
\item{nb_result}{A vector contains the AIC based on negative binomial
model for individual genes.}
\item{lognormal_result}{A vector contains the AIC based on lognormal
model for individual genes.}
\item{voom_result}{A vector contains the AIC based on voom transformation
for individual genes.}
\item{total_AIC}{The sum of AICs across all genes for the three models
in comparison.}
\item{min_AIC}{The number of minimum AIC across the three models in
comparison for individual genes.}
}
}
\details{
This function calculates the AIC based on lognormal distribution,
negative-binomial distribution as well as the Voom transformation.
It then compares the AICs of the three models across different genes.
}
\examples{
# expr_matrix: genes×samples; batch and group are length(samples)
compare_aic <- compute_aic(expr_matrix, batchind, groupind)
print(compare_aic["total_AIC"])
print(compare_aic["min_AIC"])

}
