% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{confound_metrics}
\alias{confound_metrics}
\title{This function combines std. pearson corr coef and cramer's V}
\usage{
confound_metrics(se, batch)
}
\arguments{
\item{se}{summarized experiment}

\item{batch}{batch variable}
}
\value{
metrics of confounding
}
\description{
This function combines std. pearson corr coef and cramer's V
}
