% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{confound_metrics}
\alias{confound_metrics}
\title{Combine std. Pearson correlation coefficient and Cramer's V}
\usage{
confound_metrics(se, batch)
}
\arguments{
\item{se}{summarized experiment}

\item{batch}{batch variable}
}
\value{
metrics of confounding
}
\description{
Combine std. Pearson correlation coefficient and Cramer's V
}
\examples{
library(scran)

se <- mockSCE()

confound_table <- BatchQC::confound_metrics(se, batch = "Mutation_Status")
confound_table
}
