% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variation.R
\name{covariate_pval_plotter}
\alias{covariate_pval_plotter}
\title{Covariate P-value Plotter
This function allows you to plot covariate p-values of explained variation}
\usage{
covariate_pval_plotter(DE_res)
}
\arguments{
\item{DE_res}{DE analysis results output from DE_analyze()}
}
\value{
List of explained variation by batch and condition
}
\description{
Covariate P-value Plotter
This function allows you to plot covariate p-values of explained variation
}
