% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{covariate_pval_plotter}
\alias{covariate_pval_plotter}
\title{Covariate P-value Plotter
This function allows you to plot covariate p-values of explained variation}
\usage{
covariate_pval_plotter(DE_res)
}
\arguments{
\item{DE_res}{DE analysis results output from DE_analyze()}
}
\value{
List of explained variation by batch and condition
}
\description{
Covariate P-value Plotter
This function allows you to plot covariate p-values of explained variation
}
\examples{
library(scran)

se <- mockSCE()

DE_res <- BatchQC::DE_analyze(se,
                              "DESeq2",
                              "Mutation_Status",
                              "Treatment",
                              "counts")

covariate_pval_plotter <- BatchQC::covariate_pval_plotter(DE_res = DE_res)

covariate_pval_plotter
}
