% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{covariate_pval_plotter}
\alias{covariate_pval_plotter}
\title{Covariate P-value Plotter
This function allows you to plot covariate p-values of explained variation}
\usage{
covariate_pval_plotter(DE_results)
}
\arguments{
\item{DE_results}{Differential Expression analysis result (a named list of
dataframes corresponding to each analysis completed with a "pvalue" column)}
}
\value{
boxplots of pvalues for each condition
}
\description{
Covariate P-value Plotter
This function allows you to plot covariate p-values of explained variation
}
