% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{cramers_v}
\alias{cramers_v}
\title{This function allows you to calculate Cramer's V}
\usage{
cramers_v(bd)
}
\arguments{
\item{bd}{batch design}
}
\value{
Cramer's V
}
\description{
This function allows you to calculate Cramer's V
}
