% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{DE_analyze}
\alias{DE_analyze}
\title{Differential Expression Analysis}
\usage{
DE_analyze(se, method, batch, conditions, assay_to_analyze)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{method}{DE analysis method option}

\item{batch}{Batch sample metadata column}

\item{conditions}{Sample metadata columns for additional analysis covariates}

\item{assay_to_analyze}{Assay for DE analysis}
}
\value{
A named list of of two matrices.

res features the DE analysis results.

volcano features a subset of the DE analysis results for plotting.
}
\description{
This function runs DE analysis on a count matrix in the se object
}
\examples{
library(scran)
se <- mockSCE()
# Normalized counts are required for analysis methods other than DESeq2
DE_res <- BatchQC::DE_analyze(se,
                             "DESeq2",
                             "Mutation_Status",
                             "Treatment",
                             "counts")
DE_res$dds


}
