% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{DE_analyze}
\alias{DE_analyze}
\title{Differential Expression Analysis}
\usage{
DE_analyze(se, method, batch, conditions, assay_to_analyze)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{method}{DE analysis method option (either 'DESeq' or 'limma')}

\item{batch}{metadata column in the se object representing batch}

\item{conditions}{metadata columns in the se object representing additional analysis covariates}

\item{assay_to_analyze}{Assay in the se object (either counts for DESeq or normalized data for limma) for DE analysis}
}
\value{
A named list containing the log2FoldChange, pvalue and adjusted pvalue (padj) for each analysis returned by DESeq or limma
}
\description{
This function runs DE analysis on a count matrix (DESeq) or a normalized log
or log-CPM matrix (limma) contained in the se object
}
