% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variation.R
\name{EV_plotter}
\alias{EV_plotter}
\title{This function allows you to plot explained variation}
\usage{
EV_plotter(se, batch, condition, assay_name)
}
\arguments{
\item{se}{Summarized experiment object}

\item{batch}{Batch covariate}

\item{condition}{Condition covariate of interest}

\item{assay_name}{Assay of choice}
}
\value{
List of explained variation by batch and condition
}
\description{
This function allows you to plot explained variation
}
\examples{
library(scran)
se <- mockSCE()
EV_boxplot <- BatchQC::EV_plotter(se, batch = "Mutation_Status",
                             condition = "Treatment", assay_name = "counts")
EV_boxplot

}
