% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainedVariation.R
\name{EV_plotter}
\alias{EV_plotter}
\title{This function allows you to plot explained variation}
\usage{
EV_plotter(se, batch, condition, assay_name)
}
\arguments{
\item{se}{Summarized experiment object}

\item{batch}{Batch covariate}

\item{condition}{Condition covariate of interest}

\item{assay_name}{Assay of choice}
}
\value{
List of explained variation by batch and condition
}
\description{
This function allows you to plot explained variation
}
