% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_variation.R
\name{EV_table}
\alias{EV_table}
\title{EV Table
Returns table with percent variation explained for specified number of genes}
\usage{
EV_table(se, batch, condition, assay_name)
}
\arguments{
\item{se}{Summarized experiment object}

\item{batch}{Batch Covariates}

\item{condition}{Condition covariate(s) of interest}

\item{assay_name}{Name of chosen assay}
}
\value{
List of explained variation by batch and condition
}
\description{
EV Table
Returns table with percent variation explained for specified number of genes
}
\examples{
library(scran)
se <- mockSCE()
EV_table <- BatchQC::EV_table(se, batch = "Mutation_Status",
                                    condition = "Treatment",
                                    assay_name = "counts")
EV_table

}
