% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{goodness_of_fit_DESeq2}
\alias{goodness_of_fit_DESeq2}
\title{This function calculates goodness-of-fit pvalues for all genes by looking at
how the NB model by DESeq2 fit the data}
\usage{
goodness_of_fit_DESeq2(se, count_matrix, condition, batch)
}
\arguments{
\item{se}{the se object where all the data is contained}

\item{count_matrix}{name of the se assay with gene expression matrix (in counts)}

\item{condition}{name of the se colData with the condition status}

\item{batch}{name of the se colData containing batch information}
}
\value{
a matrix of pvalues where each row is a gene and each column is a
level within the condition of interest
}
\description{
This function calculates goodness-of-fit pvalues for all genes by looking at
how the NB model by DESeq2 fit the data
}
\examples{
# example code
library(scran)
se <- mockSCE()
counts <- assays(se)["counts"]
condition <- se$Treatment
batch <- se$Mutation_Status
nb_results <- goodness_of_fit_DESeq2(count_matrix = counts,
  condition = condition, batch = batch)
nb_results
}
