% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{heatmap_plotter}
\alias{heatmap_plotter}
\title{This function allows you to plot a heatmap}
\usage{
heatmap_plotter(se, assay, nfeature, experiment_variable, annotation_column)
}
\arguments{
\item{se}{summarized experiment}

\item{assay}{normalized or corrected}

\item{nfeature}{number of features to display}

\item{experiment_variable}{what is the experiment variable}

\item{annotation_column}{choose column}
}
\value{
heatmap plot
}
\description{
This function allows you to plot a heatmap
}
