% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_plotter}
\alias{heatmap_plotter}
\title{This function allows you to plot a heatmap}
\usage{
heatmap_plotter(se, assay, nfeature, annotation_column)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{assay}{normalized or corrected assay}

\item{nfeature}{number of features to display}

\item{annotation_column}{choose column}
}
\value{
heatmap plot
}
\description{
This function allows you to plot a heatmap
}
\examples{
library(scran)
se <- mockSCE()
heatmaps <- BatchQC::heatmap_plotter(se,
                                 assay = "counts",
                                 nfeature = 15,
                                 annotation_column = c("Mutation_Status",
                                 "Treatment"))
correlation_heatmap <- heatmaps$correlation_heatmap
correlation_heatmap

heatmap <- heatmaps$topn_heatmap
heatmap

}
