% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log2CPM}
\alias{log2CPM}
\title{Compute log2(counts per mil reads) and library size for each sample}
\usage{
log2CPM(qcounts, lib.size = NULL)
}
\arguments{
\item{qcounts}{quantile normalized counts}

\item{lib.size}{default is colsums(qcounts)}
}
\value{
list containing log2(quantile counts per mil reads) and library sizes
}
\description{
Compute log2(counts per mil reads) and library size for each sample
}
\examples{
nbatch <- 3
ncond <- 2
npercond <- 10
data.matrix <- rnaseq_sim(ngenes=50, nbatch=nbatch, ncond=ncond, npercond=
    npercond, basemean=10000, ggstep=50, bbstep=2000, ccstep=800, 
    basedisp=100, bdispstep=-10, swvar=1000, seed=1234)
data.matrix <- as.matrix(data.matrix)
log2CPM(data.matrix)
}

