% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{nb_histogram}
\alias{nb_histogram}
\title{This function creates a histogram from the negative binomial goodness-of-fit
adjusted pvalues.}
\usage{
nb_histogram(adj_p_val_table)
}
\arguments{
\item{adj_p_val_table}{table of adjusted p-values from the nb test}
}
\value{
a histogram of the number of genes within a p-value range
}
\description{
This function creates a histogram from the negative binomial goodness-of-fit
adjusted pvalues.
}
