% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{nb_histogram}
\alias{nb_histogram}
\title{This function creates a histogram from the negative binomial goodness-of-fit
pvalues.}
\usage{
nb_histogram(p_val_table)
}
\arguments{
\item{p_val_table}{table of p-values from the nb test}
}
\value{
a histogram of the number of genes within a p-value range
}
\description{
This function creates a histogram from the negative binomial goodness-of-fit
pvalues.
}
\examples{
# example code
library(scran)
se <- mockSCE()
nb_results <- goodness_of_fit_DESeq2(se = se, count_matrix = "counts",
  condition = "Treatment", batch = "Mutation_Status")
nb_histogram(nb_results)
}
