% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{nb_proportion}
\alias{nb_proportion}
\title{This function determines the proportion of p-values below a specific value
and compares to the previously determined threshold}
\usage{
nb_proportion(
  p_val_table,
  low_pval,
  threshold,
  num_samples,
  small_sample_cutoff,
  method
)
}
\arguments{
\item{p_val_table}{table of p-values from the nb test}

\item{low_pval}{value of the p-value cut off to use in proportion}

\item{threshold}{the value to compare the proportion of p-values to for data
sets}

\item{num_samples}{the number of samples in the analysis}

\item{small_sample_cutoff}{value at which non-parametric test will be used
(considered "large sample size") vs parametric will be used (considered
"small sample size); default is 20}

\item{method}{string; method utilized for the parametric or non-parametric
test; either "DESeq2" or "edgeR"}
}
\value{
a statement about whether DESeq2 is appropriate to use for analysis
}
\description{
This function determines the proportion of p-values below a specific value
and compares to the previously determined threshold
}
