% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normalize_SE}
\alias{normalize_SE}
\title{This function allows you to Add normalized count matrix to the se object}
\usage{
normalize_SE(se, method, assay_to_normalize, output_assay_name)
}
\arguments{
\item{se}{SummarizeExperiment}

\item{method}{Normalization Method}

\item{assay_to_normalize}{Which assay use to do normalization}

\item{output_assay_name}{name of results assay}
}
\value{
a summarized experiment object with normalized assay appended.
}
\description{
This function allows you to Add normalized count matrix to the se object
}
