% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{PCA_plotter}
\alias{PCA_plotter}
\title{This function allows you to plot PCA}
\usage{
PCA_plotter(se, assay, nfeature, color, shape)
}
\arguments{
\item{se}{summarized experiment}

\item{assay}{normalized or corrected}

\item{nfeature}{number of features}

\item{color}{choose a color}

\item{shape}{choose a shape}
}
\value{
PCA plot
}
\description{
This function allows you to plot PCA
}
