% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{PCA_plotter}
\alias{PCA_plotter}
\title{This function allows you to plot PCA}
\usage{
PCA_plotter(se, nfeature, color, shape, assays, xaxisPC, yaxisPC)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{nfeature}{number of features}

\item{color}{choose a color}

\item{shape}{choose a shape}

\item{assays}{array of assay names from \code{se}}

\item{xaxisPC}{the PC to plot as the x axis}

\item{yaxisPC}{the PC to plot as the y axis}
}
\value{
PCA plot
}
\description{
This function allows you to plot PCA
}
