% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{permuted_edgeR}
\alias{permuted_edgeR}
\title{This function performs edgeR on the permuted dataset
adjusted pvalues.}
\usage{
permuted_edgeR(
  count_matrix,
  condition,
  other_variables,
  conditions_df,
  model_formula
)
}
\arguments{
\item{count_matrix}{matrix containing the data to be analyzed}

\item{condition}{a vector containing a factor of the condition of interest
(typically batch)}

\item{other_variables}{a vector of strings of other variables of interest}

\item{conditions_df}{data frame containing information for the other
variables of interest (columns in order of the other_variables vector)}

\item{model_formula}{the stat formula to be used in the DESeq analysis}
}
\value{
edgeR fit
}
\description{
This function performs edgeR on the permuted dataset
adjusted pvalues.
}
