% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kBET.R
\name{plot_kBET}
\alias{plot_kBET}
\title{kBET Rejection Plotter}
\usage{
plot_kBET(kBET_res)
}
\arguments{
\item{kBET_res}{list object output from kBET function}
}
\value{
ggplot object containing kBET rejection boxplot
}
\description{
This function generates a boxplot of observed and expected rejection rates
for the provided kBET output list object
}
