% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{process_dendrogram}
\alias{process_dendrogram}
\title{Process Dendrogram}
\usage{
process_dendrogram(se, assay, annotation_column)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay}{assay to plot}

\item{annotation_column}{sample metadata column}
}
\value{
named list of dendrogram data

dendrogram_segments is data representing segments of the dendrogram

dendrogram_ends is data representing ends of the dendrogram
}
\description{
This function processes count data for dendrogram plotting
}
\examples{
library(scran)
se <- mockSCE(20)

dends <- BatchQC::process_dendrogram(se,"counts","Treatment")

dends$dendrogram_segments

dends$dendrogram_ends
}
