% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchQC.R
\docType{data}
\name{protein_example_data}
\alias{protein_data}
\alias{protein_example_data}
\alias{protein_sample_info}
\title{Batch and Condition indicator for protein expression data}
\format{A data frame with 24 rows and 4 variables:
\describe{
    \item{Arrayname}{Array Name}
    \item{samplename}{Sample Name}
    \item{Batch}{Batch Indicator}
    \item{category}{Condition (Case vs Control) Indicator}
}}
\usage{
protein_sample_info

protein_data
}
\value{
Protein data sample info

Protein data
}
\description{
This data consists of two batches and two conditions 
corresponding to case and control for the protein expression data

This data consists of two batches and two conditions 
corresponding to case and control
}
\keyword{datasets}

