% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_sample_info}
\alias{protein_sample_info}
\title{Batch and Condition indicator for protein expression data}
\format{
A data frame with 24 rows and 4 variables:
\describe{
\item{Arrayname}{Array Name}
\item{samplename}{Sample Name}
\item{Batch}{Batch Indicator}
\item{category}{Condition (Case vs Control) Indicator}
}
}
\usage{
protein_sample_info
}
\description{
This data consists of two batches and two conditions
corresponding to case and control for the protein expression data
}
\examples{
data(protein_sample_info)
}
\keyword{datasets}
