% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.R
\name{pval_summary}
\alias{pval_summary}
\title{Returns summary table for p-values of explained variation}
\usage{
pval_summary(se, batch, condition, assay_name)
}
\arguments{
\item{se}{Summarized experiment object}

\item{batch}{Batch Covariate}

\item{condition}{Condition covariate(s) of interest}

\item{assay_name}{Name of chosen assay}
}
\value{
List of explained variation by batch and condition
}
\description{
Returns summary table for p-values of explained variation
}
