% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{pval_summary}
\alias{pval_summary}
\title{Returns summary table for p-values of explained variation}
\usage{
pval_summary(DE_res)
}
\arguments{
\item{DE_res}{DE analysis results output from DE_analyze()}
}
\value{
List of explained variation by batch and condition
}
\description{
Returns summary table for p-values of explained variation
}
\examples{
library(scran)
se <- mockSCE()
DE_res <- BatchQC::DE_analyze(se,
                             "DESeq2",
                             "Mutation_Status",
                             "Treatment",
                             "counts")
pval_summary <- BatchQC::pval_summary(DE_res = DE_res)
pval_summary

}
