% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial_check.R
\name{pvalues_all_genes}
\alias{pvalues_all_genes}
\title{This function calculates goodness-of-fit p-values for each condition level
for each gene size, and estimated NB mean}
\usage{
pvalues_all_genes(condition, size, mu_matrix, count_matrix)
}
\arguments{
\item{condition}{string; name of the se colData with the condition status}

\item{size}{numeric; an estimated size parameter of the NB distributions for
the gene}

\item{mu_matrix}{matrix;estimated mu parameter of the NB distributions for
different samples of each gene}

\item{count_matrix}{string; name of the assay with gene expression matrix
(in counts)}
}
\value{
a data frame of p-values for each gene under each biological
condition
}
\description{
This function calculates goodness-of-fit p-values for each condition level
for each gene size, and estimated NB mean
}
\keyword{internal}
