% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{rnaseq_sim}
\alias{rnaseq_sim}
\title{Generate simulated count data with batch effects for ngenes}
\usage{
rnaseq_sim(ngenes = 50, nbatch = 3, ncond = 2, npercond = 10,
    basemean = 10000, ggstep = 50, bbstep = 2000, ccstep = 800,
    basedisp = 100, bdispstep = 10, swvar = 1000, seed = 1000)
}
\arguments{
\item{ngenes}{Number of genes to simulate}

\item{nbatch}{Number of batches to simulate}

\item{ncond}{Number of conditions to simulate}

\item{npercond}{Number of samples per condition per batch to simulate}

\item{basemean}{Base mean}

\item{ggstep}{Gene to Gene step variation}

\item{bbstep}{Batch to Batch step variation}

\item{ccstep}{Condition to Condition step variation}

\item{basedisp}{Base Dispersion}

\item{bdispstep}{Batch to Batch Dispersion step variation}

\item{swvar}{Sample-wise extra variation}

\item{seed}{Random seed for reproducibility}
}
\value{
RNA Seq count data matrix
}
\description{
Generate simulated count data with batch effects for ngenes
}
\examples{
rnaseq_sim()
rnaseq_sim(ngenes=100, nbatch=5, seed=1234)
rnaseq_sim(ngenes=100, nbatch=3, ncond=2, npercond=10, basemean=10000,
    ggstep=50, bbstep=20000, ccstep=8000, basedisp=100, bdispstep=10, 
    swvar=1000, seed=1234)
}

