% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{std_pearson_corr_coef}
\alias{std_pearson_corr_coef}
\title{This function allows you to calculate a standardized pearson corr coef}
\usage{
std_pearson_corr_coef(bd)
}
\arguments{
\item{bd}{batch design}
}
\value{
standardized pearson correlation coefficient
}
\description{
This function allows you to calculate a standardized pearson corr coef
}
