% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{std_pearson_corr_coef}
\alias{std_pearson_corr_coef}
\title{Calculate a standardized Pearson correlation coefficient}
\usage{
std_pearson_corr_coef(bd)
}
\arguments{
\item{bd}{batch design}
}
\value{
standardized Pearson correlation coefficient
}
\description{
Calculate a standardized Pearson correlation coefficient
}
