% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{summarized_experiment}
\alias{summarized_experiment}
\title{This function creates a summarized experiment object from count
and metadata files uploaded by the user}
\usage{
summarized_experiment(counts, columndata)
}
\arguments{
\item{counts}{counts dataframe}

\item{columndata}{metadata dataframe}
}
\value{
a summarized experiment object
}
\description{
This function creates a summarized experiment object from count
and metadata files uploaded by the user
}
\examples{
data(protein_data)
data(protein_sample_info)
se_object <- summarized_experiment(protein_data, protein_sample_info)

}
