% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{summarized_experiment}
\alias{summarized_experiment}
\title{This function allows you to create a summarized experiment object}
\usage{
summarized_experiment(counts_path, metadata_path)
}
\arguments{
\item{counts_path}{path to counts file}

\item{metadata_path}{path to metadata file}
}
\value{
a summarized experiment object
}
\description{
This function allows you to create a summarized experiment object
}
