% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{sva_correction}
\alias{sva_correction}
\title{sva Correction
This function applies sva correction to a summarized experiment object
(implementation adapted from sva::psva)}
\usage{
sva_correction(
  se,
  assay_to_normalize,
  var_of_interest,
  covar,
  output_assay_name,
  psva = FALSE
)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_to_normalize}{string; name of assay that should be corrected}

\item{var_of_interest}{string; name of  experimental variable of interest}

\item{covar}{list; sting list  of covariates to include in sva analysis}

\item{output_assay_name}{string; name of results assay}

\item{psva}{boolean; default: FALSE. If set to TRUE and no covariate input,
psva function from the sva package will be used to remove batch effect.}
}
\value{
SE object with an added sva corrected array
}
\description{
sva Correction
This function applies sva correction to a summarized experiment object
(implementation adapted from sva::psva)
}
