% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{svaseq_correction}
\alias{svaseq_correction}
\title{svaseq Correction
This function applies sva correction to a summarized experiment object
with count based RNA-seq data}
\usage{
svaseq_correction(
  se,
  assay_to_normalize,
  var_of_interest,
  covar,
  output_assay_name,
  num_sv = FALSE
)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_to_normalize}{string; name of assay that should be corrected}

\item{var_of_interest}{string; name of  experimental variable of interest}

\item{covar}{list; sting list  of covariates to include in sva analysis}

\item{output_assay_name}{string; name of results assay}

\item{num_sv}{boolean; Default is FALSE: the number of estimated latent
factor is set to 1 for a small number of samples. If set to TRUE, svaseq
function will estimate the number of latent factors for you.}
}
\value{
SE object with an added sva corrected array
}
\description{
svaseq Correction
This function applies sva correction to a summarized experiment object
with count based RNA-seq data
}
