% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(volcano_data, slider)
}
\arguments{
\item{volcano_data}{A dataframe of expression change and p-value data}

\item{slider}{Magnitude of significance value threshold}
}
\value{
A volcano plot of expression change and significane value data
}
\description{
This function allows you to plot DE analysis results as a volcano plot
}
\examples{
library(scran)
se <- mockSCE()

DE_res <- BatchQC::DE_analyze(se,"DESeq2","Treatment","counts")

volcano_plot(DE_res$volcano, -150)
}
