% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{volcano_plot}
\alias{volcano_plot}
\title{This function allows you to plot differential expression analysis results as a volcano plot}
\usage{
volcano_plot(volcano_data, slider)
}
\arguments{
\item{volcano_data}{a dataframe of expression change and p-value data}
}
\value{
a volcano plot of expression change and p-value data
}
\description{
This function allows you to plot differential expression analysis results as a volcano plot
}
