% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(volcano_data, pslider, fcslider)
}
\arguments{
\item{volcano_data}{A dataframe of expression change and p-value data}

\item{pslider}{Magnitude of significance value threshold}

\item{fcslider}{Magnitude of expression change value threshold}
}
\value{
A volcano plot of expression change and significance value data
}
\description{
This function allows you to plot DE analysis results as a volcano plot
}
\examples{
library(scran)
se <- mockSCE()
DE_res <- BatchQC::DE_analyze(se,
                            "DESeq2",
                            "Mutation_Status",
                            "Treatment",
                            "counts")
volcano <- cbind(DESeq2::results(DE_res$dds)$log2FoldChange,
                                DESeq2::results(DE_res$dds)$pvalue)
plotly::ggplotly(volcano_plot(volcano, -150, 2))

}
