% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(DE_results, pslider, fcslider)
}
\arguments{
\item{DE_results}{a dataframes with the results of one of the DE Analysis;
must include "log2Foldchange" and "pvalue" columns}

\item{pslider}{Magnitude of significance value threshold}

\item{fcslider}{Magnitude of expression change value threshold}
}
\value{
A volcano plot of expression change and significance value data
}
\description{
This function allows you to plot DE analysis results as a volcano plot
}
